third_party/blink/web_tests/dom/mutation-event-tests/fast/forms/select/select-set-length-with-mutation-reorder-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMNodeRemoved' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeRemoved' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
Tests that setting the .length on an HTMLSelectElement works in the presence of DOM mutation listeners that reorder option elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sel.length is 4
PASS sel.length is 2
PASS sel.options.item(0).id is "a"
PASS sel.options.item(1).id is "b"
PASS successfullyParsed is true

TEST COMPLETE

a
b
